ContextVariables.jl

ContextVariables.jl is heavily inspired by contextvars in Python (see also PEP 567).

Tutorial

Context variables can be used to manage task-local states that are inherited to child tasks. Context variables can be created by @contextvar:

@contextvar cvar1           # untyped, without default
@contextvar cvar2 = 1       # typed (Int), with default
@contextvar cvar3::Int      # typed, without default

Note that running above code in REPL will throw an error because this form work only within a package namespace. To play with @contextvar in REPL, you can prefix the variable name with global:

julia> @contextvar global x;

Warning

@contextvar global should be used only for interactive exploration, quick scripting, and testing. Using @contextvar global inside packages make it impossible to work with serialization-based libraries such as Distributed.

The value of context variable can be get and set with the indexing syntax []

julia> x[] = 1;

julia> x[]
1

The value can be unset with delete!:

julia> delete!(x);

julia> x[]
ERROR: KeyError: key ContextVar(:x) not found

Use get and set! to handle context variables that may not be assigned:

julia> get(x)  # returns `nothing`

julia> set!(x, Some(1));  # equivalent to `x[] = 1`

julia> get(x)
Some(1)

julia> set!(x, nothing);  # equivalent to `delete!(x)`

This is useful to rollback a context variable without knowing the current value of it:

old = get(x)
x[] = some_value
do_something()
set!(x, old)  # rollback `x` to the previous state (may not be assigned)

with_context can be used to set multiple context variables at once, run a function in this context, and then rollback them to the original state:

julia> @contextvar global y = 1;
       @contextvar global z::Int;

julia> function demo1()
           @show x[]
           @show y[]
           @show z[]
       end;

julia> z[] = 100;

julia> with_context(demo1, x => :a, z => 0);
x[] = :a
y[] = 1
z[] = 0

julia> z[]
100

Note that with_context(f, x => nothing, ...) clears the value of x, rather than setting the value of x to nothing. Use Some(nothing) set nothing. Similar caution applies to set_context (see below).

julia> with_context(x => Some(nothing), y => nothing, z => nothing) do
           @show x[]
           @show y[]
           @show get(z)
       end
x[] = nothing
y[] = 1
get(z) = nothing

Thus,

with_context(x => Some(a), y => Some(b), z => Some(c)) do
    ...
end

can be used considered as a dynamically scoped version of

let x′ = a, y′ = b, z′ = c
    ...
end

Note that with_context(f, var => value, ...) does not rollback the context variables that are not specified by the input:

julia> with_context(x => :a) do
           z[] = 0
       end;

julia> z[]
0

Use with_context(f, nothing) to create an empty context and rollback the entire context to the state just before calling it.

julia> with_context(nothing) do
           z[] = 123
       end;

julia> z[]
0

Since setting multiple context variables at once is more efficient than setting them sequentially, set_context can be used to set multiple context variables in the current context in one go:

julia> set_context(x => 1, y => 2, z => 3);

julia> (x[], y[], z[])
(1, 2, 3)

A handle to the snapshot of the current context can be obtained with snapshot_context. It can be later restored by reset_context.

julia> x[]
1

julia> snapshot = snapshot_context();

julia> x[] = 123;

julia> reset_context(snapshot);

julia> x[]
1

Note that the context is inherited to the child task when the task is created. Thus, changes made after @async/@spawn or changes made in other tasks are not observable:

julia> function demo2()
           x0 = x[]
           x[] += 1
           (x0, x[])
       end

julia> with_context(x => 1) do
           @sync begin
               t1 = @async demo2()
               t2 = @async demo2()
               result = demo2()
               [result, fetch(t1), fetch(t2)]
           end
       end
3-element Array{Tuple{Int64,Int64},1}:
 (1, 2)
 (1, 2)
 (1, 2)

In particular, manipulating context variables using the public API is always data-race-free.

Warning

If a context variable holds a mutable value, it is a data-race to mutate the value when other threads are reading it.

@contextvar local x = [1]  # mutable value
@sync begin
    @spawn begin
        value = x[]        # not a data-race
        push!(value, 2)    # data-race
    end
    @spawn begin
        value = x[]        # not a data-race
        @show last(value)  # data-race
    end
end

Reference

ContextVariables.ContextVar — Type

ContextVar{T}

Context variable type. This is the type of the object var created by @contextvar var. This acts as a reference to the value stored in a task-local. The macro @contextvar is the only public API to construct this object.

Warning

It is unspecified if this type is concrete or not. It may be changed to an abstract type and/or include more type parameters in the future.

Base.get — Method

get(var::ContextVar{T}) -> Union{Some{T},Nothing}

Return Some(value) if value is assigned to var. Return nothing if unassigned.

Base.getindex — Method

getindex(var::ContextVar{T}) -> value::T

Return the value assigned to var. Throw a KeyError if unassigned.

ContextVariables.reset_context — Method

reset_context(snapshot::ContextSnapshot)

Rest the entire context of the current task to the state at which snapshot is obtained via snapshot_context.

ContextVariables.set! — Method

set!(var::ContextVar, Some(value))
set!(var::ContextVar, nothing)

Set the value of context variable var to value or delete it.

ContextVariables.set_context — Method

set_context(var1 => value1, var2 => value2, ...)
set_context(pairs)
set_context(snapshot::ContextSnapshot)

Equivalent to var1[] = value1, var2[] = value2, and so on. The second form expect an iterable of pairs of context variable and values. This function is more efficient than setting individual context variables sequentially. Like with_context, nothing value means to clear the context variable and Some is always unwrapped.

A "snapshot" of the context returned from snapshot_context can be specified as an input (the third form).

Examples

julia> using ContextVariables

julia> @contextvar global x
       @contextvar global y;

julia> set_context(x => 1, y => 2)

julia> x[]
1

julia> y[]
2

julia> set_context([x => 10, y => 20])

julia> x[]
10

julia> y[]
20

ContextVariables.snapshot_context — Method

snapshot_context() -> snapshot::ContextSnapshot

Get a snapshot of a context that can be passed to reset_context to rewind all changes in the context variables.

ContextVariables.with_context — Method

with_context(f, var1 => value1, var2 => value2, ...)
with_context(f, pairs)

Run f in a context with given values set to the context variables. Variables specified in this form are rolled back to the original value when with_context returns. It act like a dynamically scoped let. If nothing is passed as a value, corresponding context variable is cleared; i.e., it is unassigned or takes the default value. Use Some(value) to set value if value can be nothing.

with_context(f, nothing)

Run f in a new empty context. All variables are rewind to the original values when with_context returns.

Note that

var2[] = value2
with_context(var1 => value1) do
    @show var2[]  # shows value2
    var3[] = value3
end
@show var3[]  # shows value3

and

var2[] = value2
with_context(nothing) do
    var1[] = value1
    @show var2[]  # shows default (or throws)
    var3[] = value3
end
@show var3[]  # does not show value3

are not equivalent.

ContextVariables.@contextvar — Macro

@contextvar [local|global] var[::T] [= default]

Declare a context variable named var.

Warning

Context variables declared with global does not work with Distributed.

Examples

Top-level context variables needs to be declared in a package:

module MyPackage
@contextvar cvar1
@contextvar cvar2 = 1
@contextvar cvar3::Int
end

Context variables can be declared in local scope by using local prefix:

julia> using ContextVariables

julia> function demo()
           @contextvar local x = 1
           function f()
               x[] += 1
               return x[]
           end
           return f()
       end;

julia> demo()
2

To use @contextvar in a non-package namespace like REPL, prefix the variable with global:

julia> using ContextVariables

julia> @contextvar global X;

julia> X[] = 1;

julia> X[]
1